+++ src/main/resources/application-dev.yml
@@ -1,0 +1,8 @@
+spring:
+  datasource:
+    url: jdbc:h2:mem:devdb
+    driverClassName: org.h2.Driver
+    username: sa
+    password:
+  jpa:
+    show-sql: true
