+++ src/main/resources/application.yml
@@ -1,15 +1,9 @@
+spring:
+  profiles:
+    active: dev
+
 hibernate:
   physical_naming_strategy: org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy
-spring:
-  datasource:
-    url: jdbc:h2:mem:testdb
-    driverClassName: org.h2.Driver
-    username: sa
-    password:
-  jpa:
-    show-sql: true 
+
 jpa:
   database-platform: org.hibernate.dialect.H2Dialect
-server:
-  error:
-    include-stacktrace: NEVER
